requirements.txt
Flask==3.0.3
streamlit==1.39.0
PyPDF2==3.0.1
sentence-transformers==3.1.1
faiss-cpu==1.8.0
requests==2.32.3
python-dotenv==1.0.1
numpy==2.1.2